mmaprototype: move storeLoadSummary to its own file
