+++ pkg/kv/kvserver/allocator/mmaprototype/BUILD.bazel
@@ -13,6 +13,7 @@ go_library(
         "memo_helper.go",
         "messages.go",
         "rebalance_advisor.go",
+        "store_load_summary.go",
         "store_set.go",
         "top_k_replicas.go",
     ],
