+++ pkg/kv/kvserver/allocator/mmaprototype/load.go
@@ -191,29 +191,6 @@ type meanNodeLoad struct {
 	utilCPU     float64
 }
 
-type storeLoadSummary struct {
-	worstDim                                               LoadDimension // for logging only
-	sls                                                    loadSummary
-	nls                                                    loadSummary
-	dimSummary                                             [NumLoadDimensions]loadSummary
-	highDiskSpaceUtilization                               bool
-	maxFractionPendingIncrease, maxFractionPendingDecrease float64
-
-	loadSeqNum uint64
-}
-
-func (sls storeLoadSummary) String() string {
-	return redact.StringWithoutMarkers(sls)
-}
-
-func (sls storeLoadSummary) SafeFormat(w redact.SafePrinter, _ rune) {
-	w.Printf("(store=%v worst=%v cpu=%v writes=%v bytes=%v node=%v high_disk=%v frac_pending=%.2f,%.2f(%t))",
-		sls.sls, sls.worstDim, sls.dimSummary[CPURate], sls.dimSummary[WriteBandwidth], sls.dimSummary[ByteSize],
-		sls.nls, sls.highDiskSpaceUtilization, sls.maxFractionPendingIncrease,
-		sls.maxFractionPendingDecrease,
-		sls.maxFractionPendingIncrease < epsilon && sls.maxFractionPendingDecrease < epsilon)
-}
-
 // The allocator often needs mean load information for a set of stores. This
 // set is implied by a constraintsDisj. We also want to know the set of stores
 // that satisfy that contraintsDisj. meansForStoreSet encapsulates all of this
